requirements.txt
discord.py
feedparser
aiohttp